add RoomCard component
